fix: Stack-Navigation-cleanCode-types
