+++ src/routes/paramsList/AuthParamList.ts
@@ -1,0 +1,12 @@
+import { RouteProp } from "@react-navigation/native";
+import { NativeStackNavigationProp } from "@react-navigation/native-stack";
+
+export type AuthParamList = {
+    Login: undefined;
+    Signup: undefined;
+};
+  
+export type AuthNavProps<T extends keyof AuthParamList> = {
+    navigation: NativeStackNavigationProp<AuthParamList, T>
+    route:  RouteProp<AuthParamList,T>
+};
